+++ http4k-security-oauth/src/main/kotlin/org/http4k/security/openid/RequestJwts.kt
@@ -4,7 +4,7 @@ import org.http4k.security.State
 import org.http4k.security.oauth.server.AuthRequest
 
 interface RequestJwts {
-    fun create(authRequest: AuthRequest, state: State): RequestJwtContainer
+    fun create(authRequest: AuthRequest, state: State, nonce: Nonce?): RequestJwtContainer
 }
 
 data class RequestJwtContainer(val value: String)