+++ http4k-testing-servirtium/src/test/kotlin/org/http4k/servirtium/FakeGitHub.kt
@@ -6,6 +6,7 @@ import org.http4k.core.HttpHandler
 import org.http4k.core.Method.GET
 import org.http4k.core.Request
 import org.http4k.core.Response
+import org.http4k.core.Status.Companion.NOT_FOUND
 import org.http4k.core.Status.Companion.OK
 import org.http4k.core.then
 import org.http4k.core.with
@@ -13,16 +14,19 @@ import org.http4k.filter.ServerFilters
 import org.http4k.format.Jackson
 import org.http4k.format.Jackson.json
 import org.http4k.routing.bind
+import org.http4k.routing.path
 import org.http4k.routing.routes
 
 class FakeGitHub : HttpHandler {
-    private val repoFile = Jackson.parse(String(javaClass.getResourceAsStream("/org/http4k/servirtium/gitHubFile.json").readAllBytes()))
     private val app =
         ServerFilters.BasicAuth("", credentials)
             .then(
                 routes(
-                    "/repos/{owner}/{repo}/contents/{resource:.*}" bind GET to {
-                        Response(OK).with(Body.json().toLens() of repoFile)
+                    "/repos/{owner}/{repo}/contents/{resource:.+}" bind GET to {
+                        javaClass.getResourceAsStream("/" + it.path("resource") ?: "")?.let {
+                            val repoFile = Jackson.parse(String(it.readAllBytes()))
+                            Response(OK).with(Body.json().toLens() of repoFile)
+                        } ?: Response(NOT_FOUND)
                     }
                 )
             )