+++ http4k-testing-servirtium/src/test/kotlin/org/http4k/servirtium/storageProviderTests.kt
@@ -11,6 +11,7 @@ import org.http4k.testing.ApprovalTest
 import org.http4k.testing.Approver
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
+import java.nio.file.Path
 
 @ExtendWith(ApprovalTest::class)
 class GitHubStorageProviderTest {
@@ -18,20 +19,21 @@ class GitHubStorageProviderTest {
     @Test
     fun `can load content from github`(approver: Approver) {
         val storage = gitHubFor(FakeGitHub.credentials)
-        approver.assertApproved(Response(OK).body(String(storage("file.md").get())))
+        approver.assertApproved(Response(OK).body(String(storage("gitHubFile.json").get())))
     }
 
     @Test
     fun `wrong credentials fails`() {
         val storage = gitHubFor(Credentials("notuser", "noway"))
-        assertThat({ storage("file.md").get() }, throws<Unauthorized>())
+        assertThat({ storage("gitHubFile.json").get() }, throws<Unauthorized>())
     }
 
     private fun gitHubFor(credentials: Credentials) = InteractionStorage.Github(
         "owner",
         "repo",
-        "master",
         credentials,
+        Path.of("org/http4k/servirtium"),
+        "master",
         Uri.of("https://api.github.com"),
         FakeGitHub()
     )