+++ src/docs/cookbook/service_virtualisation/example_junit_contract_tests.kt
@@ -26,6 +26,7 @@ import org.junit.jupiter.api.TestInstance
 import org.junit.jupiter.api.TestInstance.Lifecycle.PER_CLASS
 import org.junit.jupiter.api.extension.RegisterExtension
 import java.io.File
+import java.nio.file.Path
 
 /**
  * This client wraps the calls to a remote WordCounter service
@@ -129,6 +130,9 @@ class ReplayFromGitHubTest : WordCounterContract {
     @JvmField
     @RegisterExtension
     val replay = ServirtiumReplay("WordCounter",
-        InteractionStorage.Github("http4k", "http4k", "master", Credentials("user", "password"))
+        InteractionStorage.Github("http4k", "http4k",
+            Credentials("user", "password"),
+            Path.of("src/test/resources/cookbook/service_virtualisation")
+        )
     )
 }