+++ CHANGELOG.md
@@ -5,6 +5,8 @@ changes with their rationale when appropriate:
 
 ### v3.228.0 (uncut)
 - [http4k-testing-servirtium] [Breaking] API is still in beta, so moving to a more composed approach which will increase reuse and allow for running Servirtium infra without a dependency on http4k or Junit.. :)
+- [http4k-security-oauth] [Breaking] Audience on request object is now a list to support multiple audiences. H/T @tom
+- [http4k-security-oauth] Nonce is now also passed through on RequestJwts, so it can be added to request jwts. H/T @tom
 
 ### v3.227.0
 - [http4k-core] Implmement #340. Support SameSite cookies. H/T @danielwellman for the contribution.